pyproject.toml
[tool.poetry]
name = "discord-auto-message-github"
version = "0.1.0"
description = ""
authors = ["ndoan <nhatminh.doan@canal-plus.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi_source"
url = "https://pypi.org/simple/"
default = true